dbeaver/pro#9798 moves user helpers to compareUser file
