+++ webapp/packages/core-authentication/src/UsersResource.ts
@@ -35,14 +35,12 @@ import { Executor } from '@cloudbeaver/core-executor';
 import { AUTH_PROVIDER_LOCAL_ID } from './AUTH_PROVIDER_LOCAL_ID.js';
 import { AuthInfoService } from './AuthInfoService.js';
 import { AuthProviderService } from './AuthProviderService.js';
+import { compareUsers, isNewUser, NEW_USER_SYMBOL, type AdminUserNew } from './compareUser.js';
 import type { IAuthCredentials } from './IAuthCredentials.js';
-
-const NEW_USER_SYMBOL = Symbol('new-user');
 
 export type AdminUser = AdminUserInfoFragment;
 export type AdminUserOrigin = AdminUserInfoFragment['origins'][number];
 
-type AdminUserNew = AdminUser & { [NEW_USER_SYMBOL]: boolean; createdAt: number };
 export type UserResourceIncludes = Omit<GetUsersListQueryVariables, 'userId' | 'page' | 'filter'>;
 
 interface IUserResourceFilterOptions {
@@ -327,45 +325,4 @@ export class UsersResource extends CachedMapResource<string, AdminUser, UserReso
   protected validateKey(key: string): boolean {
     return typeof key === 'string';
   }
-}
-
-export function isLocalUser(user: AdminUser): boolean {
-  return user.origins.some(origin => origin.type === AUTH_PROVIDER_LOCAL_ID);
-}
-
-export function isNewUser(user: AdminUser | AdminUserNew): user is AdminUserNew {
-  return NEW_USER_SYMBOL in user && user[NEW_USER_SYMBOL] === true && 'createdAt' in user && Boolean(user.createdAt);
-}
-
-export function isUser(user: object): user is AdminUser {
-  return !!user && 'userId' in user && 'grantedTeams' in user && 'enabled' in user;
-}
-
-export function compareUsers<T extends Pick<AdminUser, 'userId'>>(a: T, b: T): number {
-  return a.userId.localeCompare(b.userId);
-}
-
-export function compareUsersByLastLogin<T extends Pick<AdminUser, 'lastLoginTime'>>(a: T, b: T): number {
-  const aTime = a.lastLoginTime ? new Date(a.lastLoginTime).getTime() : 0;
-  const bTime = b.lastLoginTime ? new Date(b.lastLoginTime).getTime() : 0;
-  return aTime - bTime;
-}
-
-export function compareNewUsers(a: AdminUser, b: AdminUser): number {
-  const aIsNew = isNewUser(a);
-  const bIsNew = isNewUser(b);
-
-  if (aIsNew && !bIsNew) {
-    return -1;
-  }
-
-  if (!aIsNew && bIsNew) {
-    return 1;
-  }
-
-  if (aIsNew && bIsNew) {
-    return b.createdAt - a.createdAt;
-  }
-
-  return 0;
 }
